refactor(drawer): update imports
