+++ FinEdgeFrontend/src/Components/Notifications/LatestUnreadNotification.tsx
@@ -1,15 +1,10 @@
 import { Fade, Notification as KendoNotification, NotificationGroup } from "@progress/kendo-react-all";
-import useGetLatestNotification from "../../Hooks/useGetLatestUnreadNotification";
+import useGetLatestNotification from "../../Hooks/Notifications/useGetLatestUnreadNotification";
 import { getEnumValueFromNumber } from "../../Utils/Functions";
 import { AppNotification, NotificationType } from "../../Utils/Types";
 import { useEffect, useState } from "react";
-import axios, { AxiosError } from "axios";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { markNotificationAsReadAsync } from "../../endpoints";
 
 const LatestUnreadNotification = () => {
-  const queryClient = useQueryClient();
-
   const { data, isLoading, isError, error } = useGetLatestNotification();
   const [notification, setNotification] = useState<AppNotification | null>(null);
 
@@ -19,24 +14,10 @@ const LatestUnreadNotification = () => {
     }
   }, [data]);
 
-  const updateNotification = async () => {
-    await axios
-      .put(`${markNotificationAsReadAsync}/${notification?.id}`, { withCredentials: true })
-      .then(() => setNotification(null))
-      .catch((error: AxiosError) => {});
+  const handleClose = () => {
+    setNotification(null);
   }
 
-  const { mutateAsync } = useMutation({
-    mutationFn: updateNotification,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["notifications"] });
-    },
-  });
-
-  const handleClose = async () => {
-    await mutateAsync();
-  }
-  
   if (isLoading) return null;
   if (isError) return <p>Error: {error!.message}</p>;
 
@@ -57,7 +38,7 @@ const LatestUnreadNotification = () => {
               }}
               closable={true}
               onClose={handleClose}>
-              <span>{notification.message}</span>
+              <span>{notification.title}</span>
             </KendoNotification>
           </Fade>
         )}
